requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-search-results>=2.4.2
apscheduler>=3.10.0
pyyaml>=6.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.9
resend>=2.0.0
playwright>=1.40.0
